move common functions to scico.example
